Pull the soon-to-be-dead code to a separate file for easy removal later.
