+++ charmstore.go
@@ -210,30 +210,6 @@ func verifyHash384AndSize(path, expectHash string, expectSize int64) error {
 	return nil
 }
 
-// Latest returns the most current revision for each of the identified
-// charms. The revision in the provided charm URLs is ignored.
-func (s *CharmStore) Latest(curls ...*charm.URL) ([]CharmRevision, error) {
-	results, err := s.client.Latest(curls)
-	if err != nil {
-		return nil, err
-	}
-
-	var responses []CharmRevision
-	for i, result := range results {
-		response := CharmRevision{
-			Revision: result.Revision,
-			Sha256:   result.Sha256,
-			Err:      result.Err,
-		}
-		if errgo.Cause(result.Err) == params.ErrNotFound {
-			curl := curls[i].WithRevision(-1)
-			response.Err = CharmNotFound(curl.String())
-		}
-		responses = append(responses, response)
-	}
-	return responses, nil
-}
-
 // Resolve implements Interface.Resolve.
 func (s *CharmStore) Resolve(ref *charm.URL) (*charm.URL, []string, error) {
 	var result struct {
@@ -255,34 +231,4 @@ func (s *CharmStore) Resolve(ref *charm.URL) (*charm.URL, []string, error) {
 		return nil, nil, errgo.NoteMask(err, fmt.Sprintf("cannot resolve charm URL %q", ref), errgo.Any)
 	}
 	return result.Id.Id, result.SupportedSeries.SupportedSeries, nil
-}
-
-// URL returns the root endpoint URL of the charm store.
-func (s *CharmStore) URL() string {
-	return s.client.ServerURL()
-}
-
-// WithTestMode returns a repository Interface where test mode is enabled,
-// meaning charm store download stats are not increased when charms are
-// retrieved.
-func (s *CharmStore) WithTestMode() *CharmStore {
-	newRepo := *s
-	newRepo.client.DisableStats()
-	return &newRepo
-}
-
-// JujuMetadataHTTPHeader is the HTTP header name used to send Juju metadata
-// attributes to the charm store.
-const JujuMetadataHTTPHeader = csclient.JujuMetadataHTTPHeader
-
-// WithJujuAttrs returns a repository Interface with the Juju metadata
-// attributes set.
-func (s *CharmStore) WithJujuAttrs(attrs map[string]string) *CharmStore {
-	newRepo := *s
-	header := make(http.Header)
-	for k, v := range attrs {
-		header.Add(JujuMetadataHTTPHeader, k+"="+v)
-	}
-	newRepo.client.SetHTTPHeader(header)
-	return &newRepo
 }
